requirements.txt
PyYAML>=5.0
mistune>=2.0
PySide6>=6.0
openpyxl>=3.1.0
